requirements.txt
torch
scikit-learn
transformers
datasets
pymongo[srv]
fastapi
pydantic
motor
bson
mutagen
python-dotenv
uvicorn